+++ src/rex.deploy/src/htsql_rex_deploy/tr/signature.py
@@ -132,3 +132,7 @@ class JSONGetSig(BinarySig):
     pass
 
 
+class JSONGetJSONSig(BinarySig):
+    pass
+
+